v1.0.0.15
Fixed bug when dumping timestamps and target was defined was
indexnumber, that displayed timestamps of parent instead of target.
Fixed bug when $MFT itself contained an $ATTRIBUTE_LIST. Fixed bug with
handling records that was split across dataruns. Increased
initialization size of array for runs and vcns which caused it to crash
with extreme fragmentation.